+++ scripts/deploy.sh
@@ -12,6 +12,7 @@ cp ./canister_ids.json ./.dfx/local/canister_ids.json;
 
 # Deploy all canisters
 echo "3️⃣ Deploying all canisters..."
+dfx deploy --specified-id ryjl3-tyaaa-aaaaa-aaaba-cai icp_ledger
 dfx deploy
 
 # # Wait and verify deployment completed